# backend/schemas/__init__.py
from pydantic import BaseModel
from typing import Optional, Dict, List, Any
from datetime import datetime

class LoginRequest(BaseModel):
    email: str
    password: str

class RegisterRequest(BaseModel):
    name: str
    email: str
    password: str
    departments: List[str] = []

class OceanSubmission(BaseModel):
    name: Optional[str] = None
    openness: int
    conscientiousness: int
    extraversion: int
    agreeableness: int
    neuroticism: int

class UserProfile(BaseModel):
    id: str
    name: str
    character_class: str  
    level: int
    
    ocean_scores: Dict[str, int]  
    
    access_token: Optional[str] = None

class SkillItem(BaseModel):
    name: str
    level: int

class UserPublic(BaseModel):
    id: str
    name: str
    email: Optional[str] = None
    role: str
    character_class: str
    level: int
    ocean_openness: int
    ocean_conscientiousness: int
    ocean_extraversion: int
    ocean_agreeableness: int
    ocean_neuroticism: int
    is_available: bool
    skills: List[SkillItem] = []
    active_project_end_date: Optional[datetime] = None

class UserCandidate(BaseModel):
    id: str
    name: str
    character_class: str
    level: int
    is_available: bool
    skills: List[SkillItem] = []
    # No sensitive OCEAN scores or email

class UserListResponse(BaseModel):
    users: List[UserPublic]
    total: int

class AuthResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    user: UserPublic

class UserNameUpdate(BaseModel):
    name: str

class RoleUpdate(BaseModel):
    role: str

class MatchRequest(BaseModel):
    user1_id: str
    user2_id: str

class TeamBuilderRequest(BaseModel):
    leader_id: str
    member_count: int
    strategy: str 

class ConfirmTeamRequest(BaseModel): 
    team_name: Optional[str] = None 
    member_ids: Optional[List[str]] = None 
    start_date: datetime 
    end_date: datetime

class TeamRecommendation(BaseModel):
    strategy: str
    team_name: str
    reason: str
    leader: Dict[str, Any]
    members: List[Dict[str, Any]]
    team_score: Optional[int] = None
    team_rating: Optional[str] = None

class ReviveRequest(BaseModel):
    start_date: datetime
    end_date: datetime

class UpdateSkillsRequest(BaseModel):
    skills: List[SkillItem]

class CreateQuestRequest(BaseModel):
    prompt: str
    deadline_days: int = 7
    leader_id: str
    start_date: datetime
    deadline: datetime

class QuestSkill(BaseModel):
    name: str
    level: int

# Accepted Member Structure for Quest Response
class QuestMember(BaseModel):
    id: str
    name: str
    character_class: str
    department: str
    level: int
    matching_skills: List[Dict[str, Any]] = []

class QuestResponse(BaseModel):
    id: str
    title: str
    description: str
    rank: str
    required_skills: List[QuestSkill]
    ocean_preference: Dict[str, Any]
    team_size: int
    leader_id: str
    leader_name: Optional[str] = None
    leader_class: Optional[str] = None
    status: str
    applicant_count: int = 0
    accepted_members: List[QuestMember] = []
    accepted_member_ids: List[str] = []
    start_date: Optional[datetime] = None
    deadline: Optional[datetime] = None
    created_at: datetime
    harmony_score: Optional[int] = 0

class QuestListResponse(BaseModel):
    quests: List[QuestResponse]

class MatchScoreResponse(BaseModel):
    skill_score: int
    ocean_score: int
    total_score: int
    match_level: str
    missing_skills: List[str]
    skill_gaps: List[Dict[str, Any]]

class UpdateStatusRequest(BaseModel):
    status: str
    user_id: str

# Smart Quest (Quest 2) Schemas
class SmartQuestRequirement(BaseModel):
    department_id: str
    count: int

class PreviewSmartTeamRequest(BaseModel):
    requirements: List[SmartQuestRequirement]
    candidate_ids: List[str]

class ConfirmSmartTeamRequest(BaseModel):
    title: str
    description: Optional[str] = ""
    deadline: datetime
    start_date: datetime
    leader_id: str
    requirements: List[SmartQuestRequirement]
    member_ids: List[str]
    status: str

class AnalyzeTeamRequest(BaseModel):
    score: int
    avg_o: float
    avg_c: float
    avg_e: float
    avg_a: float
    avg_n: float